pyproject.toml
[dependency-groups]
dev = [
    "mypy>=1.17.1",
    "pre-commit>=4.3.0",
    "pytest-cov>=6.2.1",
    "pytest-env>=1.1.5",
    "pytest-mock>=3.14.1",
    "pytest>=8.4.1",
    "ruff>=0.12.11",
]

[project]
authors = [
    "Mack Nixon <mack.nixon@ucl.ac.uk>",
]
dependencies = [
    "ipython>=9.7.0",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "pyaml>=25.7.0",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.3",
    "seaborn>=0.13.2",
    "tabpfn>=6.0.6",
]

description = "A series of numerical simulations that show the (obvious) relation between personalisation and reidentification"
name = "info-gov-war"
readme = "README.md"
requires-python = ">=3.13.5"
version = "0.1.0"

[[tool.mypy.overrides]]
ignore_missing_imports = true
module = "loguru.*"

[tool.pytest.ini_options]
addopts = ["--color=yes", "--import-mode=importlib", "--verbose"]
pythonpath = "."
testpaths = ["tests"]

[tool.ruff]
extend-exclude = ["app/migrations/versions/*.py", "docs/*.py"]
fix = true
force-exclude = true
target-version = "py312"
lint.ignore = [
    "ANN003", # kwargs type annotation required
    "ARG002", # kwargs unused
    "COM812", # trailing commas (ruff-format recommended)
    "D104", # module docstring
    "D203", # no-blank-line-before-class
    "D205", # missing-blank-line-after-summary
    "D212", # multi-line-summary-first-line
    "D403", # forcing upper-case
    "D407", # removed dashes lines under sections
    "D417", # argument description in docstring (unreliable)
    "E501", # let's keep this for now (long lines)
    "ERA001", # allow commented code -- FOR NOW!
    "FIX002", # implicit trust that TODOs are not trivial, or else they would be fixed
    "ISC001", # simplify implicit str concatenation (ruff-format recommended)
    "RUF012", # where ruff wants type annotations using ClassVar for subclasses - mypy doesnt care, so why would we?
    "TD002", # name for TO DO
    "TD003", # pedantic issue number requester in to dos
]
lint.per-file-ignores = {"tests*" = [
    "ANN001", # No return type annotation
    "ANN201", # No return type annotation
    "ARG001", # Unused function argument
    "D100", # Missing docstring in public module
    "D103", # Missing docstring in public function
    "INP001", # File is part of an implicit namespace package.
    "PLR0913", # Too many arguments in function definition
    "PLR2004", # Magic value used in comparison
    "S101", # Use of `assert` detected,
    "SLF001", # accessing private members
]}
lint.select = ["ALL"]
lint.mccabe.max-complexity = 18
lint.pep8-naming.classmethod-decorators = ["classmethod"]

[tool.tomlsort]
all = true
spaces_indent_inline_array = 4
trailing_comma_inline_array = true
overrides."project.classifiers".inline_arrays = false
overrides."tool.coverage.paths.source".inline_arrays = false
